Fixes bug that the current user's avatar can be deleted.
If the current user has a custom avatar, and delete any media attachment the
user's cistom avatar is deleted too, as the delete attachment action hook
assumes that it is always the avatar that is asked to be deleted.

This updates adds a check so it is actually the avatar that is being asked
to be deleted before continuing running the avatar delete function.